major refactor of theme_builder
