--- panel/tests/themes/__init__.py
@@ -1,1 +1,0 @@
-from . import theme_test_app
