+++ panel/tests/themes/theme_builder/test_component_viewer.py
@@ -1,0 +1,9 @@
+"""In this module we test the ComponentViewer"""
+from panel.themes.theme_builder import ComponentViewer
+
+def test_view():
+    """Can view"""
+    ComponentViewer().view()
+
+if __name__.startswith("bokeh"):
+    ComponentViewer().view().servable()
